chore(row-drag): move data source for hGrid row-drag
